requirements.txt
torch>=2.0.0
numpy>=1.19.5
matplotlib>=3.5.0
pillow>=8.3.1
opencv-python>=4.5.3 